refactor Home
